backend/requirements.txt
# Django and Django REST framework
Django==4.2.10
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.2
drf-yasg==1.21.7

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.1
dj-rest-auth==5.0.2

# Celery
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# AWS
boto3==1.34.34

# Utilities
Pillow==10.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
whitenoise==6.6.0
gunicorn==21.2.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.1

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2
setuptools==69.0.3
